chore: restructure book package
