+++ common/src/main/java/com/klikli_dev/modonomicon/book/entries/CategoryLinkBookEntry.java
@@ -12,14 +12,11 @@ import com.klikli_dev.modonomicon.book.BookCategory;
 import com.klikli_dev.modonomicon.book.error.BookErrorManager;
 import com.klikli_dev.modonomicon.client.gui.BookGuiManager;
 import com.klikli_dev.modonomicon.client.gui.book.BookAddress;
-import com.klikli_dev.modonomicon.client.gui.book.category.BookCategoryNodeScreen;
-import com.klikli_dev.modonomicon.client.gui.book.entry.BookEntryScreen;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 
 public class CategoryLinkBookEntry extends BookEntry {
