+++ common/src/main/java/com/klikli_dev/modonomicon/client/gui/BookGuiManager.java
@@ -16,12 +16,16 @@ import com.klikli_dev.modonomicon.book.error.BookErrorManager;
 import com.klikli_dev.modonomicon.bookstate.BookUnlockStateManager;
 import com.klikli_dev.modonomicon.bookstate.BookVisualStateManager;
 import com.klikli_dev.modonomicon.client.gui.book.BookAddress;
+import com.klikli_dev.modonomicon.client.gui.book.BookCategoryScreen;
 import com.klikli_dev.modonomicon.client.gui.book.BookErrorScreen;
-import com.klikli_dev.modonomicon.client.gui.book.category.*;
 import com.klikli_dev.modonomicon.client.gui.book.entry.BookEntryScreen;
-import com.klikli_dev.modonomicon.client.gui.book.parent.BookParentIndexScreen;
-import com.klikli_dev.modonomicon.client.gui.book.parent.BookParentNodeScreen;
-import com.klikli_dev.modonomicon.client.gui.book.parent.BookParentScreen;
+import com.klikli_dev.modonomicon.client.gui.book.index.BookCategoryIndexOnNodeScreen;
+import com.klikli_dev.modonomicon.client.gui.book.index.BookCategoryIndexScreen;
+import com.klikli_dev.modonomicon.client.gui.book.index.BookParentIndexScreen;
+import com.klikli_dev.modonomicon.client.gui.book.node.BookCategoryNodeScreen;
+import com.klikli_dev.modonomicon.client.gui.book.node.BookParentNodeScreen;
+import com.klikli_dev.modonomicon.client.gui.book.node.DummyBookCategoryNodeScreen;
+import com.klikli_dev.modonomicon.client.gui.book.BookParentScreen;
 import com.klikli_dev.modonomicon.data.BookDataManager;
 import com.klikli_dev.modonomicon.networking.BookEntryReadMessage;
 import com.klikli_dev.modonomicon.networking.SaveBookStateMessage;
