+++ common/src/main/java/com/klikli_dev/modonomicon/client/gui/book/button/SearchButton.java
@@ -6,7 +6,7 @@
 
 package com.klikli_dev.modonomicon.client.gui.book.button;
 
-import com.klikli_dev.modonomicon.client.gui.book.parent.BookParentNodeScreen;
+import com.klikli_dev.modonomicon.client.gui.book.node.BookParentNodeScreen;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
